Move writeToTarget
